requirements.txt
# Bare minimum
django-cms==3.0



django-classy-tags==0.4
South==0.8.4
html5lib==1.0b1
django-mptt==0.6
django-sekizai==0.7
six==1.3.0
djangocms-admin-style==0.1.2

# Optional, recommended packages
Pillow==2.0.0
django-filer==0.9.5
cmsplugin-filer==0.9.5
django-reversion==1.7